Move iterchildren helpers to 'utils', make constants.py with data paths
